split html view logic
