+++ src/index.rs
@@ -1,0 +1,15 @@
+use actix_web::{error, web, Error, HttpResponse};
+use tera::Tera;
+
+async fn index(tmpl: web::Data<tera::Tera>) -> Result<HttpResponse, Error> {
+    let ctx = tera::Context::new();
+    let view = tmpl
+        .render("index.html", &ctx)
+        .map_err(|e| error::ErrorInternalServerError(e))?;
+    Ok(HttpResponse::Ok().content_type("text/html").body(view))
+}
+
+pub fn register(config: &mut web::ServiceConfig) {
+    let templates = Tera::new("templates/**/*").unwrap();
+    config.data(templates).route("/", web::get().to(index));
+}
