+++ templates/index.html
@@ -2,38 +2,6 @@
 {% block title %}カード管理{% endblock title %}
 {% block content %}
 <div class="container">
-    <h1>カード管理</h1>
-    <p class="important">
-        サーバーに保存されているカード一覧
-    </p>
-    <a href="/card">デッキ編集</a>
-    <form class="form-horizontal" method="POST" action="/card/delete" enctype="multipart/form-data">
-        <button type="submit" class="btn btn-danger">削除</button>
-        <table class="table">
-            <thead>
-                <tr>
-                    <th scope="col">チェック</th>
-                    <th scope="col">ID</th>
-                    <th scope="col">表画像</th>
-                    <th scope="col">裏画像</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for card in cards %}
-                <tr>
-                    <th>
-                        <div class="form-check">
-                            <input name="card_id[]" value="{{ card.id }}" class="form-check-input position-static"
-                                type="checkbox">
-                        </div>
-                    </th>
-                    <th>{{ card.id }}</th>
-                    <td><a href="{{ card.face }}">{{ card.face }}</td>
-                    <td><a href="{{ card.back }}">{{ card.back }}</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </form>
+    <h1>Card Playroom 管理ページ</h1>
 </div>
 {% endblock content %}
